+++ lib/simple_https_service.dart
@@ -1,4 +1,5 @@
 export 'src/manager/timing_manager.dart';
+export 'src/network/cancel_token.dart';
 export 'src/network/enum_post_type.dart';
 export 'src/network/multipart_file_spec.dart';
 export 'src/network/util_check_url.dart';